% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_OZ.R
\name{MAP_OZ}
\alias{MAP_OZ}
\title{Map Opportunity Zones}
\usage{
MAP_OZ(DIR = ".", ADDRESS = address, ZOOM = 17, OZ_CHECK = TRUE,
  MAP = TRUE, PDF = TRUE, PDF_NAME = "MAPS.pdf", XL_OUT = FALSE,
  XL_NAME = "OZ_MAP.xlsx", API_KEY = api_key)
}
\arguments{
\item{ADDRESS}{is the vector of address/es which can be input singularly following the format:/n address="1600 Pennsylvania Avenue NW, Washington, DC, 20500"/n or it can be input as a list in a .csv or .xlsx file with a column with the addresses named "addresses" and each row seperate as in :/n addresses:/n "5339 ESSERVILLE RD, NORTON, VA, 24273"/n "1305 11th Ave SE, Hickory, NC, 28602"/n "1836 SUNDANCE CIR, Yadkinville, NC, 27055"/n "514 Peach Street, Yadkinville, NC, 27055"/n}

\item{ZOOM}{allows one to change the scale of the map from 1 (lowest resolution) to 20 (highest resolution).}

\item{OZ_CHECK}{set to TRUE causes the program to evaluate whether the property is within a CDFI OZ, FALSE skips this check.}

\item{MAP}{set to TRUE outputs map to either the screen or to PDF as described below.}

\item{PDF}{set to TRUE outputs a PDF file with the sattelite maps, one page per address. FALSE plots the maps in an R window.}

\item{PDF_NAME}{is the name of the PDF file to be exported.}

\item{XL_OUT}{set to TRUE exports an Excel file to the working directory of the addresses with Census tract id and OZ determination.}

\item{XL_NAME}{set to the name of the Excel file you wish to write, this program will over write any files of the same name.}

\item{API_KEY}{is the google API key for google maps.}
}
\description{
A function to find census key, chart addresses, and evaluates whether the address is within the CDFI opportunity zones (OZ).
To run this function you must obtain a google API key by setting up an account here (https://developers.google.com/maps/documentation/embed/get-api-key).
}
\details{
Created by Steven J. Barbeaux January 25, 2019
}
\examples{
MAP_OZ()
}
\keyword{OZ}
